# Conflicts:
#	src/flight/FlightManager.java
#	src/leg/Leg.java

+++ src/leg/Leg.java
@@ -176,7 +176,7 @@ public class Leg {
      * @param seatType Type of seating
      * @return Remaining seats on leg
      */
-    public double getRemainingSeats(String seatType) {
+    public int getRemainingSeats(String seatType) {
         if (seatType.equals("First Class")) return plane.firstClassSeats() - reservedFirstClassSeats;
         else return plane.coachSeats() - reservedCoachSeats;
     }
